[refactor] GenreResponseParser 생성 및 parser 리팩토링
